Disable pull retry
This patch disables the default "podman pull" retry value, which is
not used by Infix.  Instead, the container wrapper script retries on
network related changes, or every 60 seconds.

As of podman v5.0.0 a '--retry=NUM' has been added to the podman
create, run, and pull commands.  However, CNI is no longer supported,
and a lot of other breaking changes have been made, eg., output of
podman inspect.  So there's a lot of work to upgrade.

   -- Joachim